backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
agno>=1.0.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
